All components were refactored to adapt new code style; ESLint added on project
