+++ .eslintrc.json
@@ -1,3 +1,29 @@
 {
-  "extends": "next/core-web-vitals"
+    "env": {
+        "browser": true,
+        "es2021": true
+    },
+    "extends": [
+        "plugin:react/recommended",
+        "standard"
+    ],
+    "parserOptions": {
+        "ecmaFeatures": {
+            "jsx": true
+        },
+        "ecmaVersion": "latest",
+        "sourceType": "module"
+    },
+    "plugins": [
+        "react"
+    ],
+    "rules": {
+        "react/react-in-jsx-scope": "off",
+        "indent": "off",
+        "semi": [2, "always"],
+        "quotes": "off",
+        "multiline-ternary": "off",
+        "space-before-function-paren": "off",
+        "comma-dangle": "off"
+    }
 }
