+++ components/landing/Informations/index.jsx
@@ -1,21 +1,19 @@
 import useTranslation from 'next-translate/useTranslation';
 
-import style from './informations.module.scss'
+import style from './informations.module.scss';
 
-
-const Informations = () => {
-
+export function Informations() {
     const { t } = useTranslation('home');
 
-    return(
+    return (
         <section className={style.content}>
             <article className={style.item}>
                 <div className={style.item__texts}>
                     <h2>{t('text1')}</h2>
                     <p>{t('subitle1')}</p>
                 </div>
-                <img 
-                    className={style.item__image} 
+                <img
+                    className={style.item__image}
                     src='./assets/fireplace-alt.svg'
                 />
             </article>
@@ -24,8 +22,8 @@ const Informations = () => {
                     <h2>{t('text2')}</h2>
                     <p>{t('subitle2')}</p>
                 </div>
-                <img 
-                    className={style.item__image} 
+                <img
+                    className={style.item__image}
                     src='./assets/typewriter.svg'
                 />
             </article>
@@ -34,13 +32,11 @@ const Informations = () => {
                     <h2>{t('text3')}</h2>
                     <p>{t('subitle3')}</p>
                 </div>
-                <img 
-                    className={style.item__image} 
+                <img
+                    className={style.item__image}
                     src='./assets/book.svg'
                 />
             </article>
         </section>
-    )
-}
-
-export default Informations;
+    );
+};
