+++ package.json
@@ -12,6 +12,7 @@
     "moment": "^2.29.4",
     "next": "12.0.4",
     "next-translate": "^1.2.0",
+    "prop-types": "^15.8.1",
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "react-hook-form": "^7.33.0",
@@ -20,7 +21,12 @@
     "yup": "^0.32.11"
   },
   "devDependencies": {
-    "eslint": "7.32.0",
-    "eslint-config-next": "12.0.4"
+    "eslint": "^8.0.1",
+    "eslint-config-next": "12.0.4",
+    "eslint-config-standard": "^17.0.0",
+    "eslint-plugin-import": "^2.25.2",
+    "eslint-plugin-n": "^15.0.0",
+    "eslint-plugin-promise": "^6.0.0",
+    "eslint-plugin-react": "^7.30.1"
   }
 }
