--- pages/feed/index.js
@@ -1,7 +1,0 @@
-function Feed() {
-    return(
-        <h1> Feed </h1>
-    )
-}
-
-export default Feed;
