+++ pages/feed/index.jsx
@@ -1,0 +1,7 @@
+function Feed() {
+    return (
+        <h1>Feed</h1>
+    );
+}
+
+export default Feed;
