+++ pages/index.jsx
@@ -1,25 +1,25 @@
-import Logo from '../components/design/Logo'
-import Main from '../layouts/Main';
-
-import style from '../styles/home.module.scss'
-import ButtonSection from '../components/landing/ButtonSection';
-import Informations from '../components/Landing/Informations';
-import { Button } from '../components/common/Button';
 import useTranslation from 'next-translate/useTranslation';
 
-export default function Home() {
+import { Logo } from '../components/design/Logo';
+import { ButtonSection } from '../components/landing/ButtonSection';
+import { Informations } from '../components/Landing/Informations';
+import { Button } from '../components/common/Button';
+import { Main } from '../layouts/Main';
 
+import style from '../styles/home.module.scss';
+
+function Home() {
     const { t } = useTranslation('home');
 
     return (
         <Main title='Fireplace'>
             <div className={style.content}>
-                <Button 
-                    className={style.login_button} 
-                    redirectTo='/login' 
+                <Button
+                    className={style.login_button}
+                    redirectTo='/login'
                     rightIcon='login'
                     text={t('login')}
-                />  
+                />
                 <div className={style.content__main}>
                     <div className={style.main__card}>
                         <Logo />
@@ -32,5 +32,7 @@ export default function Home() {
                 </div>
             </div>
         </Main>
-    )
+    );
 }
+
+export default Home;
