+++ pages/login/index.jsx
@@ -1,58 +1,54 @@
-import useTranslation from 'next-translate/useTranslation'
+import useTranslation from 'next-translate/useTranslation';
 import { useForm, Controller } from 'react-hook-form';
 import { useState } from 'react';
 
-import style from './login.module.scss'
-import Main from '../../layouts/Main';
-import Logo from '../../components/design/Logo';
+import style from './login.module.scss';
+import { Main } from '../../layouts/Main';
+import { Logo } from '../../components/design/Logo';
 import { Button } from '../../components/common/Button';
 import { InputText } from '../../components/common/InputText';
-import Checkbox from '../../components/common/Checkbox';
-
-
+import { Checkbox } from '../../components/common/Checkbox';
 
 function Login() {
-
     const { t } = useTranslation('login');
 
-    const [remenber, setRemember] = useState(false);
+    const [remember, setRemember] = useState(false);
 
     const { handleSubmit, control, setError, formState: { errors } } = useForm();
 
     const onToggleCheck = (check) => {
         setRemember(check);
-    }
+    };
 
     const onSubmit = async (data) => {
-        
         // --Implementation--
         //
-        //const email_exist = await api.post(data.email);
-        //if(email_exist){
+        // const email_exist = await api.post(data.email);
+        // if(email_exist){
         //      try{
         //          auth.login(data.email, data.password);
         //      } catch {
         //          setError('password', {message: t('passwordError')});
         //      }
-        //} else {
+        // } else {
         //      setError('email', {message: t('emailerror')});
-        //}
+        // }
 
-        if(data.email != 'brunosdsj@gmail.com'){
-            setError('email', {message: t('emailerror')});
-        }
-        else{
-            if(data.password != '12345678'){
-                setError('password', {message: t('passwordError')});
-            }
-            else{
+        if (data.email !== 'brunosdsj@gmail.com') {
+            setError('email', { message: t('emailerror') });
+        } else {
+            if (data.password !== '12345678') {
+                setError('password', { message: t('passwordError') });
+            } else {
+                if (remember) {
+                    console.log('lembrou');
+                }
                 console.log("Success");
             }
         }
+    };
 
-    }
-
-    return(
+    return (
         <Main
             title={t('login')}
         >
@@ -60,15 +56,15 @@ function Login() {
 
                 <Logo />
 
-                <form 
+                <form
                     className={style.form}
                     onSubmit={handleSubmit(onSubmit)}
                 >
 
-                    <Controller 
+                    <Controller
                         control={control}
                         name='email'
-                        render={ ({field:{ onChange }} ) => (
+                        render={ ({ field: { onChange } }) => (
                             <InputText
                                 type='email'
                                 variant='minimalist'
@@ -77,13 +73,13 @@ function Login() {
                                 error={errors?.email}
                             />
                         )}
-                        rules={{required: t('emailerror') }}
+                        rules={{ required: t('emailerror') }}
                     />
 
-                    <Controller 
+                    <Controller
                         control={control}
                         name='password'
-                        render={ ({field:{ onChange }}) => (
+                        render={ ({ field: { onChange } }) => (
                             <InputText
                                 type='password'
                                 variant='minimalist'
@@ -92,8 +88,8 @@ function Login() {
                                 error={errors?.password}
                             />
                         )}
-                        rules={{required: t('passwordRequired') }}
-                    />                    
+                        rules={{ required: t('passwordRequired') }}
+                    />
 
                     <Button
                         type='submit'
@@ -130,6 +126,7 @@ function Login() {
             </div>
 
         </Main>
-    )
+    );
 }
+
 export default Login;
